backend/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
pyjwt
sqlalchemy
python-multipart
